steps of transaction are slots now
